+++ ChangeLog.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Filesystem::chown()
 - Filesystem::chmod()
 - Filesystem::utime()
+- Filesystem::write()
 
 ## [0.0.3] - 2024-10-25